Complete working implementation

+++ sycl/include/sycl/__spirv/spirv_ops.hpp
@@ -197,7 +197,7 @@ extern __DPCPP_SYCL_EXTERNAL RetT __spirv_SampledImageArrayFetch(ImageT,
                                                                  TempArgT, int);
 
 template <class RetT, typename ImageT, typename TempArgT>
-extern __DPCPP_SYCL_EXTERNAL RetT __spirv_SampledImageGather(ImageT, TempArgT, unsigned long);
+extern __DPCPP_SYCL_EXTERNAL RetT __spirv_SampledImageGather(ImageT, TempArgT, int);
 
 template <class RetT, typename ImageT, typename TempArgT>
 extern __DPCPP_SYCL_EXTERNAL RetT __spirv_ImageArrayRead(ImageT, TempArgT, int);
